csvsource, sinkfunction, and source function implementation updation

+++ twister2/data/src/main/java/edu/iu/dsc/tws/data/arrow/Twister2ArrowFileWriter.java
@@ -17,13 +17,17 @@ import java.io.IOException;
 import java.util.Random;
 import java.util.logging.Logger;
 
+import com.google.common.collect.ImmutableList;
+
 import org.apache.arrow.memory.RootAllocator;
 import org.apache.arrow.vector.FieldVector;
 import org.apache.arrow.vector.IntVector;
 import org.apache.arrow.vector.VectorSchemaRoot;
 import org.apache.arrow.vector.dictionary.DictionaryProvider;
 import org.apache.arrow.vector.ipc.ArrowFileWriter;
+import org.apache.arrow.vector.types.pojo.ArrowType;
 import org.apache.arrow.vector.types.pojo.Field;
+import org.apache.arrow.vector.types.pojo.FieldType;
 import org.apache.arrow.vector.types.pojo.Schema;
 
 public class Twister2ArrowFileWriter implements ITwister2ArrowFileWriter {
@@ -31,6 +35,9 @@ public class Twister2ArrowFileWriter implements ITwister2ArrowFileWriter {
   private static final Logger LOG = Logger.getLogger(Twister2ArrowFileWriter.class.getName());
 
   private String arrowFile;
+
+  private int totalitems = 0;
+  private int totaldataValues;
 
   private int entries;
   private int maxEntries;
@@ -45,55 +52,44 @@ public class Twister2ArrowFileWriter implements ITwister2ArrowFileWriter {
 
   private Random random;
 
-  private transient FileOutputStream fileOutputStream;
+  private Twister2ArrowOutputStream twister2ArrowOutputStream;
+  private FileOutputStream fileOutputStream;
 
-  private transient RootAllocator rootAllocator = null;
+  private transient RootAllocator rootAllocator;
   private transient VectorSchemaRoot root;
   private transient ArrowFileWriter arrowFileWriter;
-
-  private int totaldataValues;
-
-  private transient Twister2ArrowOutputStream twister2ArrowOutputStream;
-
   private transient Schema arrowSchema;
 
-  private int workerId;
-  private int totalitems = 0;
-
-  public Twister2ArrowFileWriter(String arrowfile, boolean flag, Schema schema, int workerid) {
+  public Twister2ArrowFileWriter(String arrowfile, boolean flag, Schema schema) {
     this.maxEntries = 1024;
     this.checkSum = 0;
     this.arrowFile = arrowfile;
     this.flag = flag;
-    this.workerId = workerid;
-    this.arrowSchema = schema;
+    this.arrowSchema = makeSchema();
     this.rootAllocator = new RootAllocator(Integer.MAX_VALUE);
-    LOG.info("arrow file:" + this.arrowFile);
+//    try {
+//      this.fileOutputStream = new FileOutputStream(arrowFile);
+//    } catch (FileNotFoundException e) {
+//      e.printStackTrace();
+//    }
+//    LOG.info("arrow schema:" + this.arrowSchema);
+//    this.root = VectorSchemaRoot.create(this.arrowSchema, this.rootAllocator);
   }
 
-  public Twister2ArrowFileWriter(String arrowfile, boolean flag) {
-    this.maxEntries = 1024;
-    this.checkSum = 0;
-    this.batchSize = 100;
-    this.random = new Random(System.nanoTime());
-    this.entries = this.random.nextInt(this.maxEntries);
-    this.arrowFile = arrowfile;
-    this.flag = flag;
-    this.data = new int[this.entries];
-    for (int i = 0; i < this.entries; i++) {
-      this.data[i] = this.random.nextInt(1024);
-    }
-    this.rootAllocator = new RootAllocator(Integer.MAX_VALUE);
+  public Schema makeSchema() {
+    ImmutableList.Builder<Field> builder = ImmutableList.builder();
+    builder.add(new Field("int", FieldType.nullable(new ArrowType.Int(32, true)), null));
+    return new Schema(builder.build(), null);
   }
 
-  public boolean setUpTwister2ArrowWrite() throws Exception {
+  public boolean setUpTwister2ArrowWrite(int workerId) throws Exception {
+    LOG.info("arrow schema is:" + arrowSchema);
     File file = new File(arrowFile + workerId);
-    /*if (file.exists()) {
+    if (file.exists()) {
       file.delete();
-    }*/
+    }
     LOG.info("file name:" + file.getName());
     this.fileOutputStream = new FileOutputStream(arrowFile);
-    LOG.info("arrow schema:" + this.arrowSchema);
     this.root = VectorSchemaRoot.create(this.arrowSchema, this.rootAllocator);
     DictionaryProvider.MapDictionaryProvider provider
         = new DictionaryProvider.MapDictionaryProvider();
@@ -104,12 +100,13 @@ public class Twister2ArrowFileWriter implements ITwister2ArrowFileWriter {
       this.twister2ArrowOutputStream = new Twister2ArrowOutputStream(this.fileOutputStream);
       this.arrowFileWriter = new ArrowFileWriter(root, provider, this.twister2ArrowOutputStream);
     }
+    writeArrowData();
     return true;
   }
 
   public void writeArrowData(Integer integerdata) throws Exception {
     arrowFileWriter.start();
-    LOG.info("integer data:" + integerdata);
+    int i = 0;
     for (Field field : root.getSchema().getFields()) {
       FieldVector fieldVector = root.getVector(field.getName());
       IntVector intVector = (IntVector) fieldVector;
@@ -121,12 +118,7 @@ public class Twister2ArrowFileWriter implements ITwister2ArrowFileWriter {
     }
     root.setRowCount(totaldataValues);
     arrowFileWriter.writeBatch();
-//    arrowFileWriter.end();
-//    arrowFileWriter.close();
-//    fileOutputStream.flush();
-//    fileOutputStream.close();
   }
-
 
   public void writeArrowData() throws Exception {
     arrowFileWriter.start();
